app/monitoring/s3_exporters/requirements.txt
boto3 
prometheus_client 
pandas 
requests 
pytz
python-dotenv